QVAC-23075 fix: size the aspect-preserving refine in double, not float
calc_size_preserved_ratio scaled in float32 while both reference processors do it in Python
floats. Where the true product lands exactly on a multiple of align_size, float32 rounding
pushes it just past and the ceil buys a whole extra row or column of slices. 960x720 with
image_size 512 and longest_edge 2048 is the common 4:3 case: the reference refines to
2048x1536 and slices 4x3, float32 gave 2048x2048 and sliced 4x4, so the image gained 4
slices and 256 image tokens the model was never trained to see.

Verified against the local Metal build on the base checkpoint: 960x720 drops from 17 image
encodes to 13, that is a 4x4 grid to a 4x3 grid plus the overview. 1024x768, 640x480 and
256x256 are unchanged, their scale factors are exact in float32. Sweeping every 16-pixel
size pair up to 3000x3000, float32 and double disagree on 10 of 33856, all of them 4:3 or
3:4. The idefics3 and pixtral paths share this helper and both references are also double,
so they move the same way.

+++ tools/mtmd/mtmd-image.cpp
@@ -144,13 +144,18 @@ struct img_tool {
             return {0, 0};
         }
 
-        float scale = std::min(static_cast<float>(longest_edge) / inp_size.width,
-                               static_cast<float>(longest_edge) / inp_size.height);
+        // double, not float: the reference processors do this in Python floats, and where the
+        // true product lands on a multiple of align_size, float32 rounding pushes it just past
+        // and the ceil buys a whole extra row or column of slices. 960x720 at align 512 and
+        // longest_edge 2048 is the common case: 1536 in double, 2048 in float32, so a 4x3 grid
+        // becomes 4x4 and the image gains 4 slices the reference never emits.
+        double scale = std::min(static_cast<double>(longest_edge) / inp_size.width,
+                                static_cast<double>(longest_edge) / inp_size.height);
 
-        float target_width_f  = static_cast<float>(inp_size.width)  * scale;
-        float target_height_f = static_cast<float>(inp_size.height) * scale;
+        double target_width_f  = static_cast<double>(inp_size.width)  * scale;
+        double target_height_f = static_cast<double>(inp_size.height) * scale;
 
-        auto ceil_by_factor = [f = align_size](float x) { return static_cast<int>(std::ceil(x / static_cast<float>(f))) * f; };
+        auto ceil_by_factor = [f = align_size](double x) { return static_cast<int>(std::ceil(x / static_cast<double>(f))) * f; };
         int aligned_width  = ceil_by_factor(target_width_f);
         int aligned_height = ceil_by_factor(target_height_f);
